backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
claude-agent-sdk>=0.1.20
pydantic>=2.0.0
python-dotenv>=1.0.0
python-docx>=1.1.0
aiofiles>=23.0.0
e2b-code-interpreter>=1.0.0
firebase-admin>=6.5.0
